kube

+++ .zshrc
@@ -104,12 +104,56 @@ export PATH="/usr/local/sbin:/usr/local/opt/node@10/bin:/usr/local/opt/python@3.
 
 alias gg='git log --oneline --graph --all --decorate'
 alias dc='docker-compose'
+alias ip='curl ipinfo.io'
+alias markdowns='code ~/Documents/markdowns/*'
+alias timestamp='date -u +"%Y-%m-%dT%H:%M:%SZ"'
 
 dc-recompile() {
-	for service in "$@"; do
-		docker-compose stop "$service"
-	done
-	docker container prune -f
-	docker volume prune -f
-	dc up -d
+  if [ -z "$1" ]; then
+    echo "Please add the services you wish to recompile"
+  else
+    for service in "$@"
+    do
+  	docker-compose stop "$service"
+    done
+    docker container prune -f
+    docker volume prune -f
+    dc up -d
+  fi
 }
+
+k() {
+  kubectl config current-context
+}
+
+k-set() {
+  kubectl config use-context $1
+}
+
+k-get() {
+  kubectl get $1 -o wide
+}
+
+k-redeploy() {
+  kubectl patch deployment $1 -p "{\"spec\":{\"template\":{\"metadata\":{\"annotations\":{\"deploy-datetime\":\"`timestamp`\"}}}}}"
+}
+
+k-exec() {
+  echo "Finding \e[32m$1\e[39m pod..."
+  running_pods=$(kubectl get pods -l app=$1 --output=jsonpath={.items..metadata.name})
+  first_pod=$(echo $running_pods | cut -d' ' -f 1)
+  echo "Executing command...opening bash shell in pod \e[32m$first_pod \e[39m"
+  kubectl exec -it $first_pod bash
+}
+
+# tabtab source for serverless package
+# uninstall by removing these lines or running `tabtab uninstall serverless`
+[ -f /usr/local/lib/node_modules/serverless/node_modules/tabtab/.completions/serverless.bash ] && . /usr/local/lib/node_modules/serverless/node_modules/tabtab/.completions/serverless.bash
+
+# tabtab source for sls package
+# uninstall by removing these lines or running `tabtab uninstall sls`
+[ -f /usr/local/lib/node_modules/serverless/node_modules/tabtab/.completions/sls.bash ] && . /usr/local/lib/node_modules/serverless/node_modules/tabtab/.completions/sls.bash
+
+# tabtab source for slss package
+# uninstall by removing these lines or running `tabtab uninstall slss`
+[ -f /usr/local/lib/node_modules/serverless/node_modules/tabtab/.completions/slss.bash ] && . /usr/local/lib/node_modules/serverless/node_modules/tabtab/.completions/slss.bash
